requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.0
qdrant-client==1.11.1
llama-index-core==0.13.3
llama-index-readers-file==0.5.2
llama-index-vector-stores-qdrant==0.8.1
llama-index-llms-openai==0.5.4
llama-index-embeddings-openai==0.5.0
openai==1.102.0
httpx==0.27.0
pydantic==2.11.5
orjson==3.10.4
pytest==8.2.2
